+++ digital-twin-model/src/lib.rs
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 // SPDX-License-Identifier: MIT
 
+pub mod sdv_v0;
 pub mod sdv_v1;
-pub mod sdv_v2;
 
 use serde_derive::{Deserialize, Serialize};
