+++ samples/command/provider/src/main.rs
@@ -4,7 +4,7 @@
 
 mod provider_impl;
 
-use digital_twin_model::sdv_v1 as sdv;
+use digital_twin_model::sdv_v0 as sdv;
 use env_logger::{Builder, Target};
 use log::{debug, info, LevelFilter};
 use samples_common::constants::{digital_twin_operation, digital_twin_protocol};
