+++ samples/managed_subscribe/consumer/src/main.rs
@@ -4,7 +4,7 @@
 
 use std::env;
 
-use digital_twin_model::sdv_v1 as sdv;
+use digital_twin_model::sdv_v0 as sdv;
 use env_logger::{Builder, Target};
 use log::{debug, info, LevelFilter};
 use paho_mqtt as mqtt;
