+++ samples/seat_massager/consumer/src/main.rs
@@ -4,7 +4,7 @@
 
 mod respond_impl;
 
-use digital_twin_model::sdv_v2 as sdv;
+use digital_twin_model::sdv_v1 as sdv;
 use env_logger::{Builder, Target};
 use log::{debug, error, info, warn, LevelFilter};
 use rand::rngs::StdRng;
