+++ samples/seat_massager/provider/src/main.rs
@@ -4,7 +4,7 @@
 
 mod request_impl;
 
-use digital_twin_model::sdv_v2 as sdv;
+use digital_twin_model::sdv_v1 as sdv;
 use env_logger::{Builder, Target};
 use log::{debug, info, LevelFilter};
 use parking_lot::Mutex;
