+++ samples/tutorial/provider/src/main.rs
@@ -5,7 +5,7 @@
 mod provider_impl;
 
 use crate::provider_impl::ProviderImpl;
-use digital_twin_model::sdv_v1 as sdv;
+use digital_twin_model::sdv_v0 as sdv;
 use env_logger::{Builder, Target};
 use log::{info, LevelFilter};
 use samples_common::constants::{digital_twin_operation, digital_twin_protocol};
